fixes for cl_khr_command_buffer_mutable_dispatch tests
mutable_dispatch_global_size was failing because the update_global_size = 3
was not a multiple of the Local WS (original command's LWS = GWS / 64 and PoCL
doesn't support non-uniform WGs). Fixed by increasing the GWS to 256K and LWS
to 16K. The test should really detect the max Local WS of the device and use
multiples of that value - that ensures changing global WS without changing
local WS will still result in an acceptable GWS/LWS combination.

Fixes another test which had an out-of-bounds access because it hardcoded
old value of GWS.